update schema.sql and add more database code
